Architecture changes GameCollectionRow
